pyproject.toml
[project]
name = "music-player"
version = "0.1.0"
description = "A music player application with PyQt6"
authors = [
    {name = "Developer"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "pyqt6 (>=6.8.1,<7.0.0)",
    "groq (>=0.22.0,<0.23.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "qt-material (>=2.14,<3.0)",
    "qtawesome (>=1.4.0,<2.0.0)",
    "pyyaml (>=6.0.1,<7.0.0)",
    "python-vlc (>=3.0.21203,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
